requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
django==5.2; python_version >= '3.10'
django-filter==25.1; python_version >= '3.9'
djangorestframework==3.16.0; python_version >= '3.9'
drf-spectacular==0.28.0; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
markdown==3.7; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
psycopg2==2.9.10; python_version >= '3.8'
python-dotenv==1.1.0; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.24.0; python_version >= '3.9'
saxonche==12.5.0; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.13.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
werkzeug==3.1.3; python_version >= '3.9'
